Factor out files operations.
